Updated openstack/openstack Project: openstack/nova  a5cc0be5c658974a4dd2b792e23c381fd8961e23
Allow empty CPU info of hypervisors in API response

CPU info of hypervisors in an API response was changed from string to
object at API microversion 2.28. CPU info in a hypervisor object can be
empty when using some virt drivers such as ironic virt driver. If the
CPU info is an empty string or None, it cannot be convert to a JSON
object with json.loads method. It causes an internal server error.

This patch sets CPU info in an API response to an empty dict without
using json.loads method in this case follwing the schema of the
response data.

This patch also adds unit tests for this fix. These tests copies a
hypervisor object locally and modified the copy. In
CellHypervisorsTestV21, hypervisor objects are wrapped by
ComputeNodeProxy. In order to apply deepcopy to a ComputeNodeProxy,
the test adds __deepcopy__ method to the class. Without __deepcopy__,
__deepcopy__ of an self._obj works due to __getattr__
method of the proxy, which creates a copy of self._obj.

Change-Id: I34a7937ee6df985587fa98ccc3a17c969516d5ef
Closes-Bug: #1594230