Updated openstack/openstack Project: openstack/python-neutronclient  dddfaa8a3e0a5240119b5d31317040255c029539
Remove case dependancy for user inputs

With the introduction of Lowercase/Uppercase conversion
functions in the utils.py, the following patch removes the
dependancy of the current code on character casing.

utils.convert_to_lowercase is called where the code expects
the input to be in lower case while utils.convert_to_uppercase
will be called where the code expects the input to be in
CAPITAL casing only.

Change-Id: I1c5c3c87f343fc2731469b9a0c38d278f6018a11