Updated openstack/openstack Project: openstack-infra/shade  229f3daae0b9a8aba6e41eedf90eae20fcb8ecf5
Poll for image to be ready for PUT protocol

When we use PUT to upload images, they are not actually immediately
available all the time. They can come back in a "queued" state. We need
to wait until they are ready before we return the final image. Also,
sometimes the get_image call can return None, so trap for that.

Change-Id: I5b92018b848d811148377aa0fd9881d5b5daf4c7