Updated openstack/openstack Project: openstack/neutron-vpnaas  c26a4f665af9561a16d9a6beaa64b14b57fa1a09
Use temporary directory for neutron install

Do not install in /tmp/openstack/neutron and leave the git repository
there after the script is run - if zuul-cloner is used.

We run jobs on long lived workers and also on developers machines, let's
cleanup afterwards.

Install into a temporary directory that can be removed with "git clean"
afterwards. Add this directory to ignore list of pep8.

Simplify with using pushd/popd.

Also, remove setup of ZUUL_BRANCH, this is not needed with current zuul
anymore.

Sync file with neutron-lbaas.

Change-Id: Id56b9b70fb9c4113f4641e5b3f35ab63cfff6d4e