Updated openstack/openstack Project: openstack/cinder  91ee2ce67ba6750597c6d107d993d4a2ca454af4
Cleanup upgrade-related FIXME's

We've had some code paths that were related to backward compatibility
with Liberty services. As by definition we only support upgrades from
X-1 to X release, we don't need them in Newton and this commit removes
them.

This could be cleaned up earlier, but it's no harm if we'll do it now.

Related-Bug: 1609304

Change-Id: Iaaaac5a27da6104f437db3dd906099c4324f47da