Updated openstack/openstack Project: openstack/tripleo-heat-templates  649a24d1e41e3b33517e844fed794027a271a963
Rename opendaylight service to opendaylight_api

For compatibility with the custom-roles patches which convert
hard-coded hiera mappings (such as opendaylight_api_node_ips) to
data generated based on the service_name, we need to either change
this name to match the hiera (node_ips and vip) keys, or change the
hiera keys to match the service.  I took the former approach because
it involves less juggling patch dependencies between t-h-t and p-t.

Change-Id: Ic179550027f37946097afe0e5b9f504ea19fa7bc
Partially-Implements: blueprint custom-roles