Updated openstack/openstack Project: openstack/neutron-fwaas  2d09f687ed9744a65d4f62a933c76eb0aaf7dc1f
Use temporary directory for neutron install

Do not install in /tmp/openstack/neutron and leave the git repository
there after the script is run - if zuul-cloner is used.

We run jobs on long lived workers and also on developers machines, let's
cleanup afterwards.

Install into a temporary directory that can be removed with "git clean"
afterwards. Add this directory to ignore list of pep8.

Simplify with using pushd/popd which are bash features, so change
she-bang.

Also, remove setup of ZUUL_BRANCH, this is not needed with current zuul
anymore.

Move over some changes from neutron-lbaas so that these two are in sync.

Change-Id: I894cf28910c0969a8f7920f5b34c34ca0d5658b2