Updated openstack/openstack Project: openstack/ironic  9fc4ede1a23cc0c754a8fa482725e2f1a2370e3c
Fix pep8 on Python3.5

This patch is fixing the pep8 warning messages when running "tox -epep8"
with the python3.5 interpreter (the default on some distros, e.g Arch
Linux).

The method file_open() from common/utils.py was removed because it
wasn't being used anywhere and was violating pep8 by using file() which
is only present in python2.

Change-Id: Ie9356a870ea7b271aa44db57accba02f52b3e948
Closes-Bug: #1617947