Updated openstack/openstack Project: openstack/neutron-vpnaas  97cc8fc111a5d63071e0d251043779e59a701712
Add debug output to tox_install.sh

Use -x to get proper debug output of the tool.

Paul pointed out that my previous change broke neutron and
neutron-vpnaas testing together in a very specific case. Looking at this
in detail, we need some more debugging output to fix the problem. This
script is not logged right now, so, enable logging with -x and then
uplaod logs with an infra change.

Change-Id: I6546d71c51eec74c7997d0da44db79624774c073