Updated openstack/openstack Project: openstack-infra/shade  923eec6444a5b3f12d5bf43a6a02495ce4e05f2e
Refactor TaskManager to be more generic

The functionality that TaskManager represents could provide value to a
lot of people if we put it into the guts of keystoneauth so that a
person could do rate-limiting on all REST operations. It would also
catch some of the "hidden" operations, like fetching a token.
Also, TaskManager is a _very_ stable set of code, so moving it a little
further away shouldn't run the risk of deep/long bug chasing.

In preparation for that, we need to make the shade/nodepool things in
it be explicitly shade/nodepool things, and the base classes actually
be generic. Doing the refactoring here first reduces the complexity of a
move. Next step would be adding this version of TaskManager to ksa, then
in a subsequent patch we can remove it from shade. If we never do that,
nothing in this should hurt anything.

Change-Id: Id5c8caa924e8abbecef26d54d225bb10e9676d5d