Updated openstack/openstack Project: openstack-infra/shade  d53973ec28c1ecc6d4e032b9e1efe158af5427be
Add submit_function method to TaskManager

TaskManager operates on Tasks - which is awesome for things like shade
and nodepool where the discrete actions embodied by a task can be
pre-conceived. However, making a task class is a pretty heavy weight
operation to require for ad-hoc tasks. Luckily, Python lets us create
classes and closures. Add a method that allows someone to pass in a
callable or the name of a callable that should be found on client.

Change-Id: I416a7077bdd3427262243d8a39b6ed6b35ae4018