Updated openstack/openstack Project: openstack-infra/shade  42eb9e307c47db1719a57b2e8820ec91c2e0da06
Change naming style of submitTask

submitTask is camel cased because it came from nodepool, where camel
case is the prevailing style and thus correct. However, in shade,
underscores are the prevailing style, so this always feels weird.

Make the method submit_task - but add an alias for submitTask to not
break any code using the old method name.

Change-Id: I782d3b914a1b66b5af20315e58568a715222fe58