Updated openstack/openstack Project: openstack/neutron  e674abbcd99f09ad136d6db20affe3435d5c8a3b
Auto allocation: ensure that networks and subnets are cleaned up

In the event of a router create failure we need to make sure that
the allocated resources are cleaned up.

When router allocation fails it is difficult for an admin to
understand the root cause. Adding the exception here will provide
a little additional information.

Closes-bug: #1617707

Change-Id: I41940bea0327ee31494bd95867d0e60d9b5a24b8