Updated openstack/openstack Project: openstack/keystone  56d9cb51f663fe00f92a1a9f75debf943592d7bc
Make token_id a required parameter in v3_to_v2_token

The v3_to_v2_token() method in keystone.token.providers.common
accepted token_id as an option parameter. This is because it was
not always passed in on validation. This commit makes token_id
a required parameter of the method and fixes its usage to always
supply it.

Change-Id: I6cc9c7a0e306dbbad61c77caa07df00fb3fa7a97