Updated openstack/openstack Project: openstack-infra/tripleo-ci  250864166cec08c52354d6aad8b398c8de986bf6
Implement scenario001 CI job

scenario001 is a proof-of-concept of how we can scale-up our CI by
spliting services on multiple CI jobs.

In the case of scenario001:
- Sahara is deployed.
- Cinder and Swift are not deployed.
- Glance uses file backend.
- Test to register an image in Sahara with the pingtest Heat template.

Also adds the testing matrix in README. The matrix will grow with more
documentation in the future.

Note for reviewers: we will have a pingtest per scenario. After some
testing I figured it was too much complex to try to maintain common bits
in a template and do nested stacks.

Change-Id: I25521a8bf8b91f7df4020de6599c4e8420e81fef