Updated openstack/openstack Project: openstack/nova  ee9324e6d01411f2d3e29c246eb88533fa631b47
Clean up instance mappings, build requests on quota failure

During the boot process there is a point where failures, most often
quota failures, cause the newly created instance to be deleted. In this
situation the BuildRequest and InstanceMapping object that were just
created should be deleted as well. If they are not then it's possible
that an instance list will return the BuildRequest.instance as a regular
instance though it should be deleted.

Change-Id: Ic2dd3bb7db3ce563a358bed03adaa37ff12c30fd
Partially-implements: bp add-buildrequest-ob