Updated openstack/openstack Project: openstack/horizon  7328f45dc7691da8bf641c7d0796b478f09a8448
Remove admin_filter_first from horizon module to avoid confusion

The admin_filter_first attribute in
'horizon.tables.views.MultiTableMixin' causes confusion since "admin"
is an OpenStack concept. The intention of this attribute is just to
provide a setting that allows views to request for a search criteria
first before loading data, not to merge concepts between 'horizon' and
'openstack_dashboard'. Removing the setting from it's original position
and permorming the check of the setting in each admin view that
whishes to implement admin-filter-first should eliminate this confusion

Change-Id: I77fb0566155e5e5b868cb86c5718ca7048d5d7d4