Updated openstack/openstack Project: openstack/neutron  0d08350b2fbea272aa7902cae71523216382d009
Add Unit Test for Foreign Keys Not Found Error

The last addition for supporting different parent objects added a new
exception. This change pretends to cover the logic required for raising
that exception, which is when the synthetic field class doesn't have
foreign keys for the caller class.

Change-Id: Iec5bd13580ff33ef7ad97fb3846205d47d425763
Partially-Implements: blueprint adopt-oslo-versioned-objects-for-db