Updated openstack/openstack Project: openstack-infra/project-config  10cdae5aaab978372735d02f90239ea528140dff
tell the release job where the LP credentials are

The script for updating launchpad bugs when a fix is included in a
release needs to know where its credential file is. This change sets the
environment variable that the script uses to get a default so it points
to the file installed by the signing node puppet module.

Change-Id: Iaa93d047da39d27e9efbb518f77bf7844e925caf
Signed-off-by: Doug Hellmann <doug@doughellmann.com>