Updated openstack/openstack Project: openstack/neutron  aae2e8dcbc33967d1f86861c898895dc2f3e3ebd
Don't catch DBError in DHCP action handler

The DHCP port action handler has been catching DBErrors
since f1b9ac5a542a3125d757094fccda80c80c6dd420, which is
well before we had the retry decorator to deal with these.
With the port action handler catching these, it means there
will not be retries on deadlocks or connection errors so
transient situations can result in a permanently broken
DHCP service for a network.

This removes the catch for DBError so the decorator can retry
the operation.

Closes-Bug: #1618216
Change-Id: I42031b481958bbfdb8f52902c294022717af7adf