Updated openstack/openstack Project: openstack/nova  d03bbbf0e28478c7076b4b133c7a87817e762f24
driver.pre_live_migration migrate_data is always an object

The driver function pre_live_migration() has only 1 caller:
ComputeManager.pre_live_migration(), which always passes a
LiveMigrateData object to the driver. This is easy to prove as
ComputeManager.pre_live_migration() explicitly checks if migrate_data
is an object, and assigns it the return value of
LiveMigrateData.detect_implementation() if not.
detect_implementation() can only return a subclass of LiveMigrateData.
It will raise an exception if it is passed None.

This change removes handling of the case where migrate_data is None or
otherwise not a LiveMigrateData object from all driver implementations
of pre_live_migration().

Change-Id: I9d46019e06dc2b59f7b2ddf7c57c333809cbc94b