Updated openstack/openstack Project: openstack/tripleo-specs  8724ce45b579c4555aff8c5b8997acc8543c669c
Specification for tripleo-lldp-validation blueprint

This spec describes the tripleo-lldp-validation from this blueprint:
https://blueprints.launchpad.net/tripleo/+spec/tripleo-lldp-validation

The feature should gather information from Swift that was gathered
during introspection, and present a report of the physical network
topology based on LLDP data. This will make deployment
troubleshooting easier, and will provide a basis for future
automation.

Change-Id: I129080d2fee0f08155afecdaf6a94a2e535d536d