Updated openstack/openstack Project: openstack/cinder  8107b8529c186865c1ffc2b6eccbec1335e295d2
Add encryptor attach/detach in utils

To create encrypted volume from image, or retype a volume
to different encryptions, Cinder needs to attach/detch encrypted
volume.

This patch is the first patch which adds encryptor attach/detach
functions in utils so that later it is called to resolve above problems.

Change-Id: I5b6eba7e5da54cb386bf597bf100db106c7a1b06
Implements: blueprint improve-encrypted-volume