Updated openstack/openstack Project: openstack-infra/project-config  cd42908e609a818a2c75dc731ceffbfe909d5d0e
net-bgpvpn : move tempest tests to the gate queue

Now that this test look stable, and now that we have few tempest
tests, it's time to move it to the check and the gate queues.

In addition of running bgpvpn Tempest tests, this test will check
that bgpvpn and bagpipe plugins are still up and running with
the master branch of neutron.

Change-Id: I2d69edac77e18eb88ecf74071474398d15ecbe36