Updated openstack/openstack Project: openstack-infra/system-config  44fcd8e8f05c559d29cd728a46afb0af2fe67d90
Add a wiki-dev.o.o server to test newer mediawiki

The wiki-dev.openstack.org server will eventually be used to test
newer versions of Mediawiki and extensions/skins. To accommodate
this, also parameterize server backups so that they don't conflict
with production (and are in fact disabled entirely for the dev
site).

Change-Id: I6505d3af87f670e71a440c76873c085d97e5b82f