Updated openstack/openstack Project: openstack/python-ironicclient  60ecb77bc4a5404b43000ebe4c143004f316a5a9
Add --node to 'openstack baremetal port list'

This adds an optional '--node' (name or UUID of a node) to the
'openstack baremetal port list' command. It provides the ability
to get a list of ports of a specified node.

The existing release note covers this:
releasenotes/notes/osc-plugin-7769f5bef627654c.yaml.

Change-Id: Ic7064dafe061c9f751ffe0da99ede609c46c9255
Partial-Bug: #1526479