Updated openstack/openstack Project: openstack-infra/system-config  1d75ea2850b33b13e4b3c95bd049c4ff83157276
Add dual stack network config for OSIC

OSIC Cloud1 has two different provider gateway networks. One is an IPv4
network, and one is dual stack with global IPv6 and private IPv4.  List
those in the config.

The routes_externally: True entry is so that we can land the change
before the new shade release and not break old code.

Change-Id: I39d71d853290f217859a060e8a2706624550a22e