Updated openstack/openstack Project: openstack/ossa  71c48b112e81bb70227603d1225501fafd9192c2
Host a complete copy of the VMT OpenPGP keys

Directly host minimally exported copies of the VMT members' OpenPGP
keys so that they can be served securely without relying on a
third-party service. These are generated with:

    ( gpg2 --fingerprint 0x$KEY_ID;gpg2 --armor --export-options \
    export-clean,export-minimal --export 0x$KEY_ID ) > 0x$KEY_ID.txt

List complete key fingerprints as the IDs for added security, and
link to lookups of key signatures on an HTTPS-served keyserver from
the SKS pool rather than relying on Launchpad. Also correct
references from GPG to OpenPGP (GnuPG is merely one implementation
of the OpenPGP protocol), and slightly reword the prose to make it
clear that we will accept unencrypted E-mail (though encryption is
encouraged) and will accept encrypted E-mail even for less sensitive
reports.

Change-Id: Icf5f12082f1f870ec6a8f8ed09f372c952637be2