Updated openstack/openstack Project: openstack-infra/storyboard-webclient  4bd0db597861f6a8c1d98608fbd1f055d4398695
Re-add note that markdown formatting is supported

We had a note saying that StoryBoard supported markdown formatting
for text! We lost it somewhere in the Great Timeline Events Rework
of '16. This puts the note back again, and adds it to the
'create new story' dialogue for good measure.

Story: 2000712
Task: 3192

Change-Id: I24f52ec0808cf171505028dd9a25b6af8d5597e7