Updated openstack/openstack Project: openstack/cinder  3861399ec44012393eb18c46952525ab3ac42bdf
Add group snapshots - db and objects

This is the third patch that implements the generic-volume-group
bluerpint. It adds database and object changes in order to support
group snapshots and create group from source. The API changes will
be added in the next patch.

This patch depends on the second patch which adds create/delete/update
groups support which was already merged:
    https://review.openstack.org/#/c/322459/

The next patch to add volume manager changes is here:
    https://review.openstack.org/#/c/361376/

Partial-Implements: blueprint generic-volume-group
Change-Id: I2d11efe38af80d2eb025afbbab1ce8e6a269f83f