Updated openstack/openstack Project: openstack-infra/project-config  6996ff011244f8acdb94199bd31b7d08b98a1e96
Revert "Switch in Fedora 24 devstack job"

This reverts commit e9c9a7f2074621ce95c5df6a2d2817a4e6978594.

Guess what ... the Fedora 24 job is not quite ready.  Further testing
has revealed that growroot isn't working properly, so it's basically a
crap-shoot as to if the host you run on has enough free space to get
through devstack.  This is why I didn't notice during experimental
runs, which usually just ran on the single RAX node where there is
enough space.

That might be OK if it just died ... but something is happening with
ansible and it just sits there until the 3 hour timeout is finished.
So this is going to hold everything up.

Change-Id: Id3b75ae6275af1451a4922b1cb3efad368676a19