Updated openstack/openstack Project: openstack/oslo.messaging  d8c6bf7b7cdf679f85a88b465aeacca6e004ef40
[kafka] invoke TypeError exception when 'listen()' method of KafkaDriver
is called

The interface of BaseDriver which is a super-class of each underlying
transport driver has been changed.
But the interface of kafka driver doens't follow up this change. So if a
user chose it as a transport driver, an exception of TypeError would be
occurred.
This change corrects the interface in kafka's driver along with the
BaseDriver's one.

Change-Id: Iedd069b7f083e2cbf377f4148411f77ad758f979
Closes-Bug: #1616755