Updated openstack/openstack Project: openstack-infra/project-config  883fba7fd86dfaac712fb3da7798ec570f281c06
Update troves CI jobs

Commit [1] updated tox.ini in Trove and took apart the earlier py27
job and made three jobs, a py27, an apiexamples and a fakemodetests
job. This change reflects that in the CI by adding gating jobs for
apiexamples and fakemodetests.

The jobs specify a node-type of ubuntu-xenial so the branch
specification is (strictly speaking) redundant as xenial won't be used
for mitaka/liberty. But, I've left this commit message here only to be
explicit about it and so I don't wonder about it three months from now
when this nugget of information passed through the colander that
claims to be my brain.

[1] https://review.openstack.org/#/c/362203

Co-Authored-By: Andreas Jaeger <aj@suse.com>
Change-Id: I69cc5ce233bd330a6d98ffe1598dea0c02a5d7e9
Depends-On: I0a971c6ef6c1a4669490f2aa0e67c795fe86e190