Updated openstack/openstack Project: openstack-infra/project-config  47f8a7a7f5fa74946181005d0f3cc27893a41043
Add k8s-docker-suite-app-murano project to openstack

As was discussed in mail thread
http://lists.openstack.org/pipermail/openstack-dev/2016-June/096458.html
the main aim of this patch to move some applications to separate
repository. It should have the same commit story and be maintained the
same team as early. Current repository contains several groups of
applications:
 - Docker applications, which deploy docker itself.
 - Kubernetes applications, for deploying Kubernetes.
 - Several applications adopted fro runing them in Kubernetes cluster
   mentioned above.

The purpose of current project is to develop and maintain applications
mentioned above. So this project is not a Murano development, it's about
development specific applications, which can be used by Murano.

Also new core group will include murano-apps-core group, because
previously this group maintained these applications.

Change-Id: Ida451ccd2c69590dcf2bfd21977cd1b854626c02